Refactor user models
