+++ Cantinarr/Features/OverseerrUsers/Logic/TrendingViewModel.swift
@@ -6,7 +6,6 @@ import SwiftUI
 /// Fetches and caches trending movies/TV shows from Overseerr.
 @MainActor
 final class TrendingViewModel: ObservableObject {
-    typealias MediaItem = OverseerrUsersViewModel.MediaItem // reuse struct
 
     // Dependencies
     private let service: OverseerrServiceType
